# docs/openapi.json
{
  "components": {
    "schemas": {
      "ErrorDetail": {
        "description": "Mirrors `platform.errors.error_envelope`'s inner `\"error\"` object --\na SEPARATE pydantic model (not imported from `platform.errors`, which\nis framework-free by design), used only to describe the shape in the\ngenerated OpenAPI document (tasks.md 7b.2).",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "`{\"error\": {code, message, details?}}` -- design.md's error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationStatus": {
        "description": "Mirrors the `validation_status` CHECK constraint (migration 0001, the\nAlembic revision that creates the `phrases` table -- see design.md's\n\"Data Model and Migrations\").",
        "enum": [
          "unique",
          "duplicate_confirmed"
        ],
        "title": "ValidationStatus",
        "type": "string"
      },
      "_MatchesData": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/_ScoredPhrase"
            },
            "title": "Matches",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "matches",
          "next_cursor",
          "has_more",
          "total"
        ],
        "title": "_MatchesData",
        "type": "object"
      },
      "_MatchesRequest": {
        "properties": {
          "cursor": {
            "description": "Opaque; clients MUST NOT parse it.",
            "title": "Cursor",
            "type": "string"
          },
          "limit": {
            "anyOf": [
              {
                "maximum": 50.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "cursor"
        ],
        "title": "_MatchesRequest",
        "type": "object"
      },
      "_MatchesResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/_MatchesData"
          }
        },
        "required": [
          "data"
        ],
        "title": "_MatchesResponse",
        "type": "object"
      },
      "_PhraseListData": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/_PhraseOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "next_cursor",
          "has_more"
        ],
        "title": "_PhraseListData",
        "type": "object"
      },
      "_PhraseListResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/_PhraseListData"
          }
        },
        "required": [
          "data"
        ],
        "title": "_PhraseListResponse",
        "type": "object"
      },
      "_PhraseOut": {
        "description": "The 201 body of `POST /phrases`.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "validation": {
            "$ref": "#/components/schemas/_ValidationOut"
          }
        },
        "required": [
          "id",
          "text",
          "created_at",
          "validation"
        ],
        "title": "_PhraseOut",
        "type": "object"
      },
      "_PhraseResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/_PhraseOut"
          }
        },
        "required": [
          "data"
        ],
        "title": "_PhraseResponse",
        "type": "object"
      },
      "_SaveRequest": {
        "properties": {
          "confirm_duplicate": {
            "default": false,
            "title": "Confirm Duplicate",
            "type": "boolean"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "_SaveRequest",
        "type": "object"
      },
      "_ScoredPhrase": {
        "description": "Shared shape of `most_similar` and each `matches[]` entry.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "score"
        ],
        "title": "_ScoredPhrase",
        "type": "object"
      },
      "_ValidateData": {
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "is_duplicate": {
            "title": "Is Duplicate",
            "type": "boolean"
          },
          "matches": {
            "items": {
              "$ref": "#/components/schemas/_ScoredPhrase"
            },
            "title": "Matches",
            "type": "array"
          },
          "most_similar": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/_ScoredPhrase"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "threshold": {
            "title": "Threshold",
            "type": "number"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "is_duplicate",
          "threshold",
          "score",
          "most_similar",
          "matches",
          "next_cursor",
          "has_more",
          "total"
        ],
        "title": "_ValidateData",
        "type": "object"
      },
      "_ValidateRequest": {
        "properties": {
          "limit": {
            "anyOf": [
              {
                "maximum": 50.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "_ValidateRequest",
        "type": "object"
      },
      "_ValidateResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/_ValidateData"
          }
        },
        "required": [
          "data"
        ],
        "title": "_ValidateResponse",
        "type": "object"
      },
      "_ValidationOut": {
        "properties": {
          "most_similar_phrase_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Most Similar Phrase Id"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "validated_at": {
            "format": "date-time",
            "title": "Validated At",
            "type": "string"
          }
        },
        "required": [
          "status",
          "score",
          "most_similar_phrase_id",
          "validated_at"
        ],
        "title": "_ValidationOut",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "get_health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Health"
      }
    },
    "/phrases": {
      "get": {
        "operationId": "list_phrases_phrases_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 200,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ValidationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "min_score",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1,
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_PhraseListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `INVALID_CURSOR`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `VALIDATION_ERROR`."
          }
        },
        "summary": "List Phrases"
      },
      "post": {
        "operationId": "save_phrase_phrases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_SaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_PhraseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `DUPLICATE_CONFIRMATION_REQUIRED`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `VALIDATION_ERROR`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `EMBEDDING_UNAVAILABLE`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `EMBEDDING_TIMEOUT`."
          }
        },
        "summary": "Save Phrase"
      }
    },
    "/phrases/matches": {
      "post": {
        "operationId": "list_matches_phrases_matches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_MatchesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_MatchesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `INVALID_CURSOR`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `VALIDATION_ERROR`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `EMBEDDING_UNAVAILABLE`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `EMBEDDING_TIMEOUT`."
          }
        },
        "summary": "List Matches"
      }
    },
    "/phrases/validate": {
      "post": {
        "operationId": "validate_phrase_phrases_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_ValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `VALIDATION_ERROR`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `EMBEDDING_UNAVAILABLE`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `EMBEDDING_TIMEOUT`."
          }
        },
        "summary": "Validate Phrase"
      }
    },
    "/phrases/{phrase_id}": {
      "get": {
        "description": "Single phrase by id -- `apps/web`'s \"compare with the matched\nphrase\" panel resolves `validation.most_similar_phrase_id` through\nthis, since the list/save/validate responses only ever carry that\nid, never the matched phrase's own text (design: avoid joining it\ninto every list row when most rows are never expanded).",
        "operationId": "get_phrase_phrases__phrase_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "phrase_id",
            "required": true,
            "schema": {
              "title": "Phrase Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/_PhraseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error envelope; `error.code` = `PHRASE_NOT_FOUND`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Phrase"
      }
    }
  }
}